http client: remove the nng_http_client_transact function
This function required the request and response to be preallocated
separately, which is a little awkward.  Instead, we have the
nng_http_conn_transact function.  This function does require a
connection to be established first, but then it is easier to manage.

This is part of overall goal of removing separate allocation and
removal of request and reply objects.